refactoring contest select for reuse
