+++ src/app/contestselect/contestselect.component.scss
