+++ src/app/genericcontent.scss
