--- src/app/team-select/team.select.component.html
@@ -1,7 +1,0 @@
-<div class="select-team-container">
-  <router-outlet></router-outlet>
-<!--  <team (sendSelections)="addTeams($event)"></team>-->
-</div>
-
-
-<!-- TODO: On click, ensure that background color changes to verify selection -->
