+++ src/app/team-select/team.select.routes.ts
@@ -2,10 +2,9 @@ import { Routes } from '@angular/router';
 
 import { TeamComponent } from './team/team.component';
 import { TeamSelectComponent } from './team.select.component';
-import { ContestSelectComponent } from './contestselect/contestselect.component';
+import { ContestSelectComponent } from '../contestselect/contestselect.component';
 export const TeamSelectRoutes: Routes = [{
     path: 'entercontest',
-    component: TeamSelectComponent,
     children: [
         {path: '', component: ContestSelectComponent},
         {path: ':year/:gender', component: TeamComponent},
