+++ src/app/updatebracket/updatebracket.component.html
@@ -5,3 +5,4 @@
     <h1>{{region.name}}</h1>
     <app-bracket [bracket]="region.bracket" [displayUntil]=64 [top]="true"></app-bracket>
 </div>
+<button mat-raised-button (click)="onSubmit()">Submit</button>
